requirements.txt
Django==2.2.10
gunicorn==20.0.04
djangorestframework==3.12.2
whitenoise==5.2.0

